refactor(SettingsContactEdit): code
